Added unit test for reading GZipped Clinvar HGVS file.
