+++ tests/unit_tests/phpunit.xml
@@ -2,6 +2,7 @@
     <testsuites>
         <testsuite name="main">
             <file>GetVariantInfoTest.php</file>
+            <file>ClinvarFileTest.php</file>
         </testsuite>
     </testsuites>
 </phpunit>
